chmod -x for unnecessary files (NO content change)
I left configure and install scripts to have executable access
but removed executable access from all other source, definition,
documentation files that are not meant to run as a script.
It's kind of annoying to see all files in green color or something
when you list the files in UNIX or Linux terminal.